+++ tsconfig.json
@@ -2,6 +2,7 @@
   "compilerOptions": {
     "target": "ES2020",
     "module": "commonjs",
+    "declaration": true,
     "outDir": "./dist",
     "rootDir": "./src",
     "strict": true,
